+++ CHANGES.rst
@@ -5,6 +5,8 @@ threedi-qgis-plugin changelog
 0.14 (unreleased)
 -----------------
 
+- Also make it possible to delete control groups from the database and from view with the 'Delete button' in the 'Control group' tab.
+
 - Also make it possible to delete rule from the database and from view with the 'Delete button' in the 'Rule' tab.
 
 - Make it possible to delete measuring group from the database and from view with the 'Delete button in the 'Measuring group' tab.